@@ -1,22 +1,21 @@
 import "styles/index.css";
 import React from "react";
-import ReactDOM from "react-dom";
-import ThemedApp from "./app/ThemedApp";
+import { createRoot } from "react-dom/client";
+import ThemedApp from "app/ThemedApp.jsx";
 import { Provider } from "react-redux";
-import store from "redux/store.js";
+import store from "store/store";
 
 import reportWebVitals from "./reportWebVitals";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <ThemedApp />
     </Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
